+++ protoc_plugin/lib/src/base_type.dart
@@ -56,6 +56,9 @@ class BaseType {
   String getRepeatedDartType(FileGenerator fileGen) =>
       '$coreImportPrefix.List<${getDartType(fileGen)}>';
 
+  String getRepeatedDartTypeIterable(FileGenerator fileGen) =>
+      '$coreImportPrefix.Iterable<${getDartType(fileGen)}>';
+
   factory BaseType(FieldDescriptorProto field, GenerationContext ctx) {
     String constSuffix;