+++ protoc_plugin/test/generated_message_test.dart
@@ -5,6 +5,10 @@
 import 'package:protobuf/protobuf.dart';
 import 'package:test/test.dart';
 
+import '../out/protos/constructor_args/google/protobuf/unittest.pb.dart'
+    as constructor_args_unittest;
+import '../out/protos/constructor_args/google/protobuf/unittest_import.pb.dart'
+    as constructor_args_unittest_import;
 import '../out/protos/duplicate_names_import.pb.dart';
 import '../out/protos/google/protobuf/unittest.pb.dart';
 import '../out/protos/google/protobuf/unittest_import.pb.dart';
@@ -783,5 +787,107 @@ void main() {
     setAllExtensions(value1);
     final value2 = value1.deepCopy();
     assertAllExtensionsSet(value2);
+  });
+
+  test('Named arguments in constructors', () {
+    final value = constructor_args_unittest.TestAllTypes(
+      optionalInt32: 101,
+      optionalInt64: make64(102),
+      optionalUint32: 103,
+      optionalUint64: make64(104),
+      optionalSint32: 105,
+      optionalSint64: make64(106),
+      optionalFixed32: 107,
+      optionalFixed64: make64(108),
+      optionalSfixed32: 109,
+      optionalSfixed64: make64(110),
+      optionalFloat: 111.0,
+      optionalDouble: 112.0,
+      optionalBool: true,
+      optionalString: '115',
+      optionalBytes: '116'.codeUnits,
+      optionalGroup:
+          constructor_args_unittest.TestAllTypes_OptionalGroup(a: 117),
+      optionalNestedMessage:
+          constructor_args_unittest.TestAllTypes_NestedMessage(bb: 118),
+      optionalForeignMessage: constructor_args_unittest.ForeignMessage(c: 119),
+      optionalImportMessage:
+          constructor_args_unittest_import.ImportMessage(d: 120),
+      optionalNestedEnum: constructor_args_unittest.TestAllTypes_NestedEnum.BAZ,
+      optionalForeignEnum: constructor_args_unittest.ForeignEnum.FOREIGN_BAZ,
+      optionalImportEnum:
+          constructor_args_unittest_import.ImportEnum.IMPORT_BAZ,
+      optionalStringPiece: '124',
+      optionalCord: '125',
+      repeatedInt32: [201, 301],
+      repeatedInt64: [make64(202), make64(302)],
+      repeatedUint32: [203, 303],
+      repeatedUint64: [make64(204), make64(304)],
+      repeatedSint32: [205, 305],
+      repeatedSint64: [make64(206), make64(306)],
+      repeatedFixed32: [207, 307],
+      repeatedFixed64: [make64(208), make64(308)],
+      repeatedSfixed32: [209, 309],
+      repeatedSfixed64: [make64(210), make64(310)],
+      repeatedFloat: [211.0, 311.0],
+      repeatedDouble: [212.0, 312.0],
+      repeatedBool: [true, false],
+      repeatedString: ['215', '315'],
+      repeatedBytes: ['216'.codeUnits, '316'.codeUnits],
+      repeatedGroup: [
+        constructor_args_unittest.TestAllTypes_RepeatedGroup(a: 217),
+        constructor_args_unittest.TestAllTypes_RepeatedGroup(a: 317)
+      ],
+      repeatedNestedMessage: [
+        constructor_args_unittest.TestAllTypes_NestedMessage(bb: 218),
+        constructor_args_unittest.TestAllTypes_NestedMessage(bb: 318)
+      ],
+      repeatedForeignMessage: [
+        constructor_args_unittest.ForeignMessage(c: 219),
+        constructor_args_unittest.ForeignMessage(c: 319)
+      ],
+      repeatedImportMessage: [
+        constructor_args_unittest_import.ImportMessage(d: 220),
+        constructor_args_unittest_import.ImportMessage(d: 320)
+      ],
+      repeatedNestedEnum: [
+        constructor_args_unittest.TestAllTypes_NestedEnum.BAR,
+        constructor_args_unittest.TestAllTypes_NestedEnum.BAZ
+      ],
+      repeatedForeignEnum: [
+        constructor_args_unittest.ForeignEnum.FOREIGN_BAR,
+        constructor_args_unittest.ForeignEnum.FOREIGN_BAZ
+      ],
+      repeatedImportEnum: [
+        constructor_args_unittest_import.ImportEnum.IMPORT_BAR,
+        constructor_args_unittest_import.ImportEnum.IMPORT_BAZ
+      ],
+      repeatedStringPiece: ['224', '324'],
+      repeatedCord: ['225', '325'],
+      defaultInt32: 401,
+      defaultInt64: make64(402),
+      defaultUint32: 403,
+      defaultUint64: make64(404),
+      defaultSint32: 405,
+      defaultSint64: make64(406),
+      defaultFixed32: 407,
+      defaultFixed64: make64(408),
+      defaultSfixed32: 409,
+      defaultSfixed64: make64(410),
+      defaultFloat: 411.0,
+      defaultDouble: 412.0,
+      defaultBool: false,
+      defaultString: '415',
+      defaultBytes: '416'.codeUnits,
+      defaultNestedEnum: constructor_args_unittest.TestAllTypes_NestedEnum.FOO,
+      defaultForeignEnum: constructor_args_unittest.ForeignEnum.FOREIGN_FOO,
+      defaultImportEnum: constructor_args_unittest_import.ImportEnum.IMPORT_FOO,
+      defaultStringPiece: '424',
+      defaultCord: '425',
+    );
+
+    // Convert the message with constructor arguments to the message without
+    // constructor arguments, to be able to reuse `assertAllFieldsSet`.
+    assertAllFieldsSet(TestAllTypes.fromBuffer(value.writeToBuffer()));
   });
 }